requirements.txt
pillow==10.3.0
playsound==1.3.0
pyobjc==10.2